Update remaining reduction ops to use MIOpen.  double datatype is not supported, so disable those typed kernels.

+++ orttraining/orttraining/training_ops/rocm/reduction/reduction_ops.cc
@@ -28,8 +28,8 @@ namespace rocm {
 
 REGISTER_MS_KERNEL_TYPED(ReduceSumTraining, MLFloat16)
 REGISTER_MS_KERNEL_TYPED(ReduceSumTraining, float)
-REGISTER_MS_KERNEL_TYPED(ReduceSumTraining, double)
-// REGISTER_MS_KERNEL_TYPED(ReduceSumTraining, int32_t)
+// REGISTER_MS_KERNEL_TYPED(ReduceSumTraining, double)
+REGISTER_MS_KERNEL_TYPED(ReduceSumTraining, int32_t)
 
 template <bool allow_multi_axes>
 template <typename T, miopenReduceTensorIndices_t ReduceTensorIndices>
